feat : DAY06 : Do World-Coord Transform

+++ Framework/Systems/Keyboard.cpp
@@ -6,16 +6,15 @@ Keyboard * Keyboard::Instance = nullptr;
 
 void Keyboard::Create()
 {
-	ASSERT(Instance == nullptr, "Keyboard Instance must null on Create()");
+	ASSERT(static_cast<int>(Instance == nullptr), "Keyboard Instance must null on Create()");
 	Instance = new Keyboard();
 
 }
 
 void Keyboard::Destroy()
 {
-	ASSERT(Instance != nullptr, "Keyboard Instance must not null on Destroy()");
+	ASSERT(static_cast<int>(Instance != nullptr), "Keyboard Instance must not null on Destroy()");
 	SAFE_DELETE(Instance);
-
 }
 
 Keyboard* Keyboard::Get()
